create stream in manager thread
This prepares for the next change, reverting to sending the request
immediately on stream creation.

If we create two streams A and B, where idB = idA+2, then we must
make sure to send A's frames first. If we send B's frames first,
then A automatically gets closed (see http2 spec).